services/proxy/requirements.txt
uvicorn
fastapi
httpx
transformers
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-fastapi
pygelf
python-dateutil
tiktoken
Jinja2